file restructuring and using hydra radius graph function
