--- examples/LennardJones/distributed_utils.py
@@ -1,3 +1,0 @@
-def nsplit(a, n):
-    k, m = divmod(len(a), n)
-    return (a[i * k + min(i, m) : (i + 1) * k + min(i + 1, m)] for i in range(n))
